✨ Add statistics page

+++ lib/ui/pages/main/project/project_details/view/project_details_page.dart
@@ -162,9 +162,15 @@ class _ProjectFullCardWidget extends StatelessWidget {
             child: LinearPercentIndicatorWidget(percent: project.percentCompleted),
           ),
           if (isMember)
-            const Padding(
-              padding: EdgeInsets.symmetric(horizontal: 24, vertical: 12),
-              child: YunoWhiteTextButton(text: 'Invite People'),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+              child: GestureDetector(
+                onTap: () async => context.pushNamed(
+                  RouteName.projectMembers,
+                  params: {'id': project.id},
+                ),
+                child: const YunoWhiteTextButton(text: 'Project Members'),
+              ),
             )
           else
             Padding(
@@ -211,7 +217,9 @@ class _ButtonsRowWidget extends StatelessWidget {
                     builder: (context) => const YunoAlertDialog(text: 'Delete The Project?'),
                   );
                   if (val ?? true) {
-                    context.read<ProjectDetailsBloc>().add(const ProjectDetailsEvent.delete());
+                    if (context.mounted) {
+                      context.read<ProjectDetailsBloc>().add(const ProjectDetailsEvent.delete());
+                    }
                   }
                 }
               : null,
@@ -248,7 +256,9 @@ class _ButtonsRowWidget extends StatelessWidget {
               builder: (context) => const YunoAlertDialog(text: 'Leave The Project?'),
             );
             if (val ?? true) {
-              context.read<ProjectDetailsBloc>().add(const ProjectDetailsEvent.leave());
+              if (context.mounted) {
+                context.read<ProjectDetailsBloc>().add(const ProjectDetailsEvent.leave());
+              }
             }
           },
           child: YunoIconButton(
